Chore: delete more unused files
