+++ frontend/src/component/project/Project/CreateProject/NewCreateProjectForm/CreateProjectDialog.styles.tsx
@@ -1,0 +1,14 @@
+import { styled } from '@mui/material';
+
+export const StyledDefinitionList = styled('dl')(({ theme }) => ({
+    dt: {
+        fontWeight: 'bold',
+        '&:after': {
+            content: '":"',
+        },
+    },
+
+    'dd + dt': {
+        marginBlockStart: theme.spacing(1),
+    },
+}));
